requirements.txt
load-dotenv==0.1.0
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dotenv==1.0.0
